Move bunch of functions to nodes.py
